move binance to separate location in source
